created constant/index.ts and added all prev used data #6
