+++ src/components/custom/Testimonial.tsx
@@ -4,6 +4,7 @@ import { GradientLight } from "../ui/Gradient";
 import Marquee from "react-fast-marquee";
 import { FaLinkedin } from "react-icons/fa";
 import useDeviceType from "../../hooks/useDeviceType";
+import {benefits} from "../../constants";
 
 const Benefits = () => {
 const isMobile = useDeviceType();
@@ -64,64 +65,3 @@ const BenefitCard = ({ item }: { item: any }) => {
 };
 
 export default Benefits;
-
-export const benefits = [
-  {
-    id: "0",
-    title: "Rishabh Jain",
-    text: "Working with Nishant has been an absolute pleasure. His expertise in the Web Technology and attention to detail have significantly improved our project's performance.",
-    backgroundUrl: " ../../public/benefits/card-1.svg",
-    iconUrl: "",
-    imageUrl: "",
-    pos:'Full Stack Developer',
-    linkedin:""
-  },
-  // {
-  //   id: "1",
-  //   title: "",
-  //   text: "The app uses natural language processing to understand user queries and provide accurate and relevant responses.",
-  //   backgroundUrl: " ../../public/benefits/card-2.svg",
-  //   iconUrl: benefitIcon2,
-  //   imageUrl: benefitImage2,
-  //   light: true,
-  //   pos:'SDE-II Google',
-  //   linkedin:""
-  // },
-  {
-    id: "2",
-    title: "Karan Mashru",
-    text: "Nishant's technical articles are well-researched and engaging. His work has significantly increased our audience engagement and provided valuable insights into emerging technologies.",
-    backgroundUrl: " ../../public/benefits/card-3.svg",
-    iconUrl: "",
-    imageUrl: "",
-    pos:'Ex - Mentor GeeksforGeeks',
-    linkedin:""
-  },
-  // {
-  //   id: "3",
-  //   title: "Fast responding",
-  //   text: "Lets users quickly find answers to their questions without having to search through multiple sources.",
-  //   backgroundUrl: " ../../public/benefits/card-4.svg",
-  //   iconUrl: benefitIcon4,
-  //   imageUrl: benefitImage2,
-  //   light: true,
-  //   pos:'SDE-II Google',
-  //   linkedin:""
-  // },
-  {
-    id: "4",
-    title: "Rahul Kapoor",
-    text: "Nishant is an innovative developer who proposed the idea for a self-learning platform for small vendors. His creativity and technical skills significantly impacted our projects. ",
-    backgroundUrl: " ../../public/benefits/card-5.svg",
-    pos:'Joint Secretary, MoHUA',
-    linkedin:""
-  },
-  {
-    id: "5",
-    title: "Dr. Deep Chandra",
-    text: "The app uses natural language processing to understand user queries and provide accurate and relevant responses.",
-    backgroundUrl: " ../../public/benefits/card-6.svg",
-    pos:'As. Professor at VIT Bhopal',
-    linkedin:""
-  },
-];
